+++ playwright/integrations/authenticated-chat.spec.ts
@@ -1,15 +1,15 @@
 import fetchOmnichannelConfig from '../utils/fetchOmnichannelConfig';
 import fetchTestPageUrl from '../utils/fetchTestPageUrl';
+import fetchAuthUrl from '../utils/fetchAuthUrl';
 import { test, expect } from '@playwright/test';
 import OmnichannelEndpoints from '../utils/OmnichannelEndpoints';
-import ACSEndpoints from '../utils/ACSEndpoints';
 
 const testPage = fetchTestPageUrl();
 const omnichannelConfig = fetchOmnichannelConfig('AuthenticatedChat');
+const authUrl = fetchAuthUrl('AuthenticatedChat');
 
 test.describe('AuthenticatedChat @AuthenticatedChat', () => {
-
-    test('Calling ChatSDK.startChat() should not fail for authenticated Chat', async ({ page }) => {
+    test('ChatSDK.startChat() should fetch the chat token & perform session init', async ({page}) => {
         await page.goto(testPage);
 
         const [chatTokenRequest, chatTokenResponse, sessionInitRequest, sessionInitResponse, runtimeContext] = await Promise.all([
@@ -25,13 +25,19 @@ test.describe('AuthenticatedChat @AuthenticatedChat', () => {
             page.waitForResponse(response => {
                 return response.url().includes(OmnichannelEndpoints.LiveChatAuthSessionInitPath);
             }),
-            await page.evaluate(async ({ omnichannelConfig }) => {
-                const { OmnichannelChatSDK_1: OmnichannelChatSDK } = window;
+            await page.evaluate(async ({ omnichannelConfig, authUrl }) => {
+                const {OmnichannelChatSDK_1: OmnichannelChatSDK} = window;
+
+                const payload = {
+                    method: "POST"
+                };
+
+                const response = await fetch(authUrl, payload);
+                const authToken = await response.text();
+
                 const chatSDKConfig = {
-                    getAuthToken: async () => {
-                        return omnichannelConfig.token;
-                    }
-                }
+                    getAuthToken: () => authToken
+                };
 
                 const chatSDK = new OmnichannelChatSDK.default(omnichannelConfig, chatSDKConfig);
 
@@ -39,6 +45,7 @@ test.describe('AuthenticatedChat @AuthenticatedChat', () => {
 
                 const runtimeContext = {
                     requestId: chatSDK.requestId,
+                    authToken
                 };
 
                 await chatSDK.startChat();
@@ -46,159 +53,21 @@ test.describe('AuthenticatedChat @AuthenticatedChat', () => {
                 await chatSDK.endChat();
 
                 return runtimeContext;
-            }, { omnichannelConfig })
+            }, { omnichannelConfig, authUrl })
         ]);
 
-        const { requestId } = runtimeContext;
+        const {requestId, authToken} = runtimeContext;
         const chatTokenRequestUrl = `${omnichannelConfig.orgUrl}/${OmnichannelEndpoints.LiveChatv2AuthGetChatTokenPath}/${omnichannelConfig.orgId}/${omnichannelConfig.widgetId}/${requestId}?channelId=lcw`;
         const sessionInitRequestUrl = `${omnichannelConfig.orgUrl}/${OmnichannelEndpoints.LiveChatAuthSessionInitPath}/${omnichannelConfig.orgId}/${omnichannelConfig.widgetId}/${requestId}?channelId=lcw`;
 
+        const chatTokenRequestHeaders = chatTokenRequest.headers();
+        const sessionInitRequestHeaders = sessionInitRequest.headers();
+
         expect(chatTokenRequest.url() === chatTokenRequestUrl).toBe(true);
+        expect(chatTokenRequestHeaders['authenticatedusertoken']).toBe(authToken);
         expect(chatTokenResponse.status()).toBe(200);
         expect(sessionInitRequest.url() === sessionInitRequestUrl).toBe(true);
+        expect(sessionInitRequestHeaders['authenticatedusertoken']).toBe(authToken);
         expect(sessionInitResponse.status()).toBe(200);
-
-        const chatrequestHeaders = chatTokenRequest.headers();
-        expect(chatrequestHeaders['authenticatedusertoken']).toBe(omnichannelConfig.token);
-        const sessioninitrequestHeaders = sessionInitRequest.headers();
-        expect(sessioninitrequestHeaders['authenticatedusertoken']).toBe(omnichannelConfig.token);
-    });
-
-    test('ChatSDK.startChat() with a liveChatContext should validate the live work item details and auth chat map record & should not perform session init call', async ({ page }) => {
-        await page.goto(testPage);
-
-        const requestUrls = [];
-        page.on('request', (request) => {
-            if (request.url().includes("livechatconnector")) {
-                requestUrls.push(request.url());
-            }
-        });
-
-        const [_, liveWorkItemDetailsRequest, liveWorkItemDetailsResponse, liveChatMapRecordRequest, liveChatMapRecordResponse, runtimeContext] = await Promise.all([
-            await page.evaluate(async ({ omnichannelConfig }) => {
-                const { OmnichannelChatSDK_1: OmnichannelChatSDK } = window;
-                const chatSDKConfig = {
-                    getAuthToken: async () => {
-                        return omnichannelConfig.token;
-                    }
-                }
-                const chatSDK = new OmnichannelChatSDK.default(omnichannelConfig, chatSDKConfig);
-
-                await chatSDK.initialize();
-
-                await chatSDK.startChat();
-
-                const liveChatContext = await chatSDK.getCurrentLiveChatContext();
-
-                const runtimeContext = {
-                    runtimeIdFirstSession: chatSDK.runtimeId,
-                    requestIdFirstSession: chatSDK.requestId,
-                    liveChatContext,
-                };
-
-                (window as any).runtimeContext = runtimeContext;
-            }, { omnichannelConfig }),
-            page.waitForRequest(request => {
-                return request.url().includes(OmnichannelEndpoints.LiveChatAuthLiveWorkItemDetailsPath);
-            }),
-            page.waitForResponse(response => {
-                return response.url().includes(OmnichannelEndpoints.LiveChatAuthLiveWorkItemDetailsPath);
-            }),
-            page.waitForRequest(request => {
-                return request.url().includes(OmnichannelEndpoints.LiveChatAuthChatMapRecord);
-            }),
-            page.waitForResponse(response => {
-                return response.url().includes(OmnichannelEndpoints.LiveChatAuthChatMapRecord);
-            }),
-            await page.evaluate(async ({ omnichannelConfig }) => {
-                const { OmnichannelChatSDK_1: OmnichannelChatSDK, runtimeContext } = window;
-
-                const chatSDKConfig = {
-                    getAuthToken: async () => {
-                        return omnichannelConfig.token;
-                    }
-                }
-
-                const chatSDK = new OmnichannelChatSDK.default(omnichannelConfig, chatSDKConfig);
-
-                await chatSDK.initialize();
-
-                runtimeContext.runtimeIdSecondSession = chatSDK.runtimeId;
-                runtimeContext.requestIdSecondSession = chatSDK.requestId;
-
-                try {
-                    await chatSDK.startChat({ liveChatContext: runtimeContext.liveChatContext });
-                } catch (err) {
-                    runtimeContext.errorMessage = `${err.message}`;
-                }
-
-                await chatSDK.endChat();
-
-                return runtimeContext;
-            }, { omnichannelConfig })
-        ]);
-
-        const { requestIdFirstSession: requestId } = runtimeContext;
-        const liveWorkItemDetailsRequestUrl = `${omnichannelConfig.orgUrl}/${OmnichannelEndpoints.LiveChatAuthLiveWorkItemDetailsPath}/${omnichannelConfig.orgId}/${omnichannelConfig.widgetId}/${requestId}?channelId=lcw`;
-        const authChatMapRecordRequestUrl = `${omnichannelConfig.orgUrl}/${OmnichannelEndpoints.LiveChatAuthChatMapRecord}/${omnichannelConfig.orgId}/${omnichannelConfig.widgetId}`;
-        const liveWorkItemDetailsResponseDataJson = await liveWorkItemDetailsResponse.json();
-        const sessionInitCalls = requestUrls.filter((url) => url.includes(OmnichannelEndpoints.LiveChatAuthSessionInitPath));
-
-        expect(liveWorkItemDetailsRequest.url() === liveWorkItemDetailsRequestUrl).toBe(true);
-        expect(liveWorkItemDetailsResponse.status()).toBe(200);
-        expect(liveChatMapRecordRequest.url()).toContain(authChatMapRecordRequestUrl);
-        expect(liveChatMapRecordResponse.status()).toBe(200);
-        expect(liveWorkItemDetailsResponseDataJson.State).not.toBe('Closed');
-        expect(sessionInitCalls.length).toBe(1);
-    });
-
-    test('ChatSDK.sendMessage() should send the message masked for authenticated chat', async ({ page }) => {
-        await page.goto(testPage);
-
-        const content = "4111111111111111";
-
-        const [getMessagesRequest, getMessagesResponse, runtimeContext] = await Promise.all([
-            page.waitForRequest(request => {
-                return request.url().match(ACSEndpoints.getMessagesPathPattern)?.length >= 0 && request.method() === 'GET';
-            }),
-            page.waitForResponse(response => {
-                return response.url().match(ACSEndpoints.getMessagesPathPattern)?.length >= 0 && response.request().method() === 'GET'
-            }),
-            await page.evaluate(async ({ omnichannelConfig, content }) => {
-                const { OmnichannelChatSDK_1: OmnichannelChatSDK } = window;
-                const chatSDKConfig = {
-                    getAuthToken: async () => {
-                        return omnichannelConfig.token;
-                    }
-                }
-                const chatSDK = new OmnichannelChatSDK.default(omnichannelConfig, chatSDKConfig);
-
-                await chatSDK.initialize();
-
-                await chatSDK.startChat();
-
-                const runtimeContext = {
-                    requestId: chatSDK.requestId,
-                    chatId: chatSDK.chatToken.chatId,
-                    acsEndpoint: chatSDK.chatToken.acsEndpoint,
-                };
-
-                await chatSDK.sendMessage({
-                    content
-                });
-
-                await chatSDK.getMessages();
-
-                await chatSDK.endChat();
-
-                return runtimeContext;
-            }, { omnichannelConfig, content })
-        ]);
-
-        const getMessagesResponseDataJson = await getMessagesResponse.json();
-        const sentMessages = getMessagesResponseDataJson.value.filter((message) => message.type === 'text');
-        const sentMessageContent = sentMessages[0].content.message;
-
-        expect(sentMessageContent).toBe("################");
     });
 });